+++ CHANGELOG.md
@@ -3,6 +3,7 @@
  - Improved Enterprise fake AP compatibility
  - Improved cleanup routine to remove temporary files on unexpected script termination
  - Modified the lighttpd web server configuration during the Evil Twin captive portal attack to facilitate integration with certain plugins
+ - Added downgrade attack against WPA2/WPA3 transitional (Mixed Mode) networks in the WPA3 menu
 
 ### 11.52
  - Add WPA3 attacks menu as hookable function to facilitate WPA3 existing plugins integration